refactor: transfer process module
