+++ app/build.gradle.kts
@@ -83,6 +83,7 @@ dependencies {
     implementation(projects.feature.guarantor)
     implementation(projects.feature.savings)
     implementation(projects.feature.qr)
+    implementation(projects.feature.transferProcess)
