+++ settings.gradle.kts
@@ -34,3 +34,4 @@ include(":feature:beneficiary")
 include(":feature:registration")
 include(":feature:savings")
 include(":feature:qr")
+include(":feature:transfer-process")
